% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isCusip}
\alias{isCusip}
\title{Check validity of CUSIP code}
\usage{
isCusip(s)
}
\arguments{
\item{s}{nine-character string with CUSIP code to validate}
}
\description{
This function check validity of CUSIP code by comparing calculated check digit based on first 8 characters of CUSIP code with 9th character of CUSIP code - control digit
}
\keyword{CUSIP}
\keyword{Control}
\keyword{Digit}
