% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isIsin}
\alias{isIsin}
\title{Check validity of ISIN code}
\usage{
isIsin(s)
}
\arguments{
\item{s}{twelve-character string with ISIN code to validate}
}
\value{
Logical value indicating validity of ISIN number: \code{TRUE} if ISIN number is valid, \code{FALSE} if ISIN number is invalid
}
\description{
This function check validity of ISIN code by comparing calculated checksum digit based on first 11 characters of CUSIP code with 12th character of CUSIP code - checksum digit
}
\examples{
library(SEC13Flist)
isIsin("US0378331009") #invalid ISIN example
isIsin("US0378331005") #valid ISIN example
}
\keyword{ISIN}
\keyword{checksum}
\keyword{digit}
