% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSedol.R
\name{isSedol}
\alias{isSedol}
\title{Check validity of SEDOL identifier}
\usage{
isSedol(s)
}
\arguments{
\item{s}{seven-character string with SEDOL code to validate}
}
\value{
Logical value indicating validity of SEDOL number: \code{TRUE} if SEDOL number is valid, \code{FALSE} if SEDOL number is invalid
}
\description{
This function check validity of SEDOL code by comparing calculated checksum digit based on first 6 characters of SEDOL code with 7th character of SEDOL code - checksum digit
}
\examples{
library(SEC13Flist)
isSedol("2046252") #invalid SEDOL example
isSedol("2046251") #valid SEDOL example
}
\keyword{SEDOL}
\keyword{checksum}
\keyword{digit}
