% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{line_separator_func}
\alias{line_separator_func}
\title{Line separator}
\usage{
line_separator_func(CR, LF, text)
}
\arguments{
\item{CR}{CR position (integer)}

\item{LF}{LF position (integer)}

\item{text}{character}
}
\description{
internal function that determines line separator for text
}
