% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{List_13F_2004_Q1}
\alias{List_13F_2004_Q1}
\alias{List_13F_2004_Q2}
\alias{List_13F_2004_Q3}
\alias{List_13F_2004_Q4}
\alias{List_13F_2005_Q1}
\alias{List_13F_2005_Q2}
\alias{List_13F_2005_Q3}
\alias{List_13F_2005_Q4}
\alias{List_13F_2006_Q1}
\alias{List_13F_2006_Q2}
\alias{List_13F_2006_Q3}
\alias{List_13F_2006_Q4}
\alias{List_13F_2007_Q1}
\alias{List_13F_2007_Q2}
\alias{List_13F_2007_Q3}
\alias{List_13F_2007_Q4}
\alias{List_13F_2008_Q1}
\alias{List_13F_2008_Q2}
\alias{List_13F_2008_Q3}
\alias{List_13F_2008_Q4}
\alias{List_13F_2009_Q1}
\alias{List_13F_2009_Q2}
\alias{List_13F_2009_Q3}
\alias{List_13F_2009_Q4}
\alias{List_13F_2010_Q1}
\alias{List_13F_2010_Q2}
\alias{List_13F_2010_Q3}
\alias{List_13F_2010_Q4}
\alias{List_13F_2011_Q1}
\alias{List_13F_2011_Q2}
\alias{List_13F_2011_Q3}
\alias{List_13F_2011_Q4}
\alias{List_13F_2012_Q1}
\alias{List_13F_2012_Q2}
\alias{List_13F_2012_Q3}
\alias{List_13F_2012_Q4}
\alias{List_13F_2013_Q1}
\alias{List_13F_2013_Q2}
\alias{List_13F_2013_Q3}
\alias{List_13F_2013_Q4}
\alias{List_13F_2014_Q1}
\alias{List_13F_2014_Q2}
\alias{List_13F_2014_Q3}
\alias{List_13F_2014_Q4}
\alias{List_13F_2015_Q1}
\alias{List_13F_2015_Q2}
\alias{List_13F_2015_Q3}
\alias{List_13F_2015_Q4}
\alias{List_13F_2016_Q1}
\alias{List_13F_2016_Q2}
\alias{List_13F_2016_Q3}
\alias{List_13F_2016_Q4}
\alias{List_13F_2017_Q1}
\alias{List_13F_2017_Q2}
\alias{List_13F_2017_Q3}
\alias{List_13F_2017_Q4}
\alias{List_13F_2023_Q1}
\alias{List_13F_2023_Q2}
\alias{List_13F_2023_Q3}
\alias{List_13F_2023_Q4}
\alias{List_13F_2024_Q1}
\title{Official List of Section 13(f) Securities historical data}
\format{
\subsection{\code{List_13F_2004_Q1}}{

A data frame with 13,541 rows and 7 columns
}

\subsection{\code{List_13F_2004_Q2}}{

A data frame with 13,683 rows and 7 columns
}

\subsection{\code{List_13F_2004_Q3}}{

A data frame with 13,799 rows and 7 columns
}

\subsection{\code{List_13F_2004_Q4}}{

A data frame with 13,965 rows and 7 columns
}

\subsection{\code{List_13F_2005_Q1}}{

A data frame with 14,056 rows and 7 columns
}

\subsection{\code{List_13F_2005_Q2}}{

A data frame with 14,139 rows and 7 columns
}

\subsection{\code{List_13F_2005_Q3}}{

A data frame with 14,266 rows and 7 columns
}

\subsection{\code{List_13F_2005_Q4}}{

A data frame with 14,296 rows and 7 columns
}

\subsection{\code{List_13F_2006_Q1}}{

A data frame with 14,294 rows and 7 columns
}

\subsection{\code{List_13F_2006_Q2}}{

A data frame with 14,375 rows and 7 columns
}

\subsection{\code{List_13F_2006_Q3}}{

A data frame with 14,407 rows and 7 columns
}

\subsection{\code{List_13F_2006_Q4}}{

A data frame with 14,822 rows and 7 columns
}

\subsection{\code{List_13F_2007_Q1}}{

A data frame with 14,844 rows and 7 columns
}

\subsection{\code{List_13F_2007_Q2}}{

A data frame with 15,246 rows and 7 columns
}

\subsection{\code{List_13F_2007_Q3}}{

A data frame with 15,460 rows and 7 columns
}

\subsection{\code{List_13F_2007_Q4}}{

A data frame with 15,510 rows and 7 columns
}

\subsection{\code{List_13F_2008_Q1}}{

A data frame with 15,532 rows and 7 columns
}

\subsection{\code{List_13F_2008_Q2}}{

A data frame with 15,670 rows and 7 columns
}

\subsection{\code{List_13F_2008_Q3}}{

A data frame with 15,645 rows and 7 columns
}

\subsection{\code{List_13F_2008_Q4}}{

A data frame with 15,621 rows and 7 columns
}

\subsection{\code{List_13F_2009_Q1}}{

A data frame with 15,383 rows and 7 columns
}

\subsection{\code{List_13F_2009_Q2}}{

A data frame with 15,262 rows and 7 columns
}

\subsection{\code{List_13F_2009_Q3}}{

A data frame with 15,216 rows and 7 columns
}

\subsection{\code{List_13F_2009_Q4}}{

A data frame with 15,319 rows and 7 columns
}

\subsection{\code{List_13F_2010_Q1}}{

A data frame with 15,383 rows and 7 columns
}

\subsection{\code{List_13F_2010_Q2}}{

A data frame with 15,262 rows and 7 columns
}

\subsection{\code{List_13F_2010_Q3}}{

A data frame with 15,216 rows and 7 columns
}

\subsection{\code{List_13F_2010_Q4}}{

A data frame with 15,319 rows and 7 columns
}

\subsection{\code{List_13F_2011_Q1}}{

A data frame with 15,684 rows and 7 columns
}

\subsection{\code{List_13F_2011_Q2}}{

A data frame with 15,889 rows and 7 columns
}

\subsection{\code{List_13F_2011_Q3}}{

A data frame with 15,974 rows and 7 columns
}

\subsection{\code{List_13F_2011_Q4}}{

A data frame with 16,010 rows and 7 columns
}

\subsection{\code{List_13F_2012_Q1}}{

A data frame with 15,965 rows and 7 columns
}

\subsection{\code{List_13F_2012_Q2}}{

A data frame with 15,997 rows and 7 columns
}

\subsection{\code{List_13F_2012_Q3}}{

A data frame with 16,112 rows and 7 columns
}

\subsection{\code{List_13F_2012_Q4}}{

A data frame with 16,137 rows and 7 columns
}

\subsection{\code{List_13F_2013_Q1}}{

A data frame with 16,080 rows and 7 columns
}

\subsection{\code{List_13F_2013_Q2}}{

A data frame with 16,247 rows and 7 columns
}

\subsection{\code{List_13F_2013_Q3}}{

A data frame with 16,339 rows and 7 columns
}

\subsection{\code{List_13F_2013_Q4}}{

A data frame with 16,532 rows and 7 columns
}

\subsection{\code{List_13F_2014_Q1}}{

A data frame with 16,648 rows and 7 columns
}

\subsection{\code{List_13F_2014_Q2}}{

A data frame with 16,784 rows and 7 columns
}

\subsection{\code{List_13F_2014_Q3}}{

A data frame with 17,072 rows and 7 columns
}

\subsection{\code{List_13F_2014_Q4}}{

A data frame with 17,235 rows and 7 columns
}

\subsection{\code{List_13F_2015_Q1}}{

A data frame with 17,229 rows and 7 columns
}

\subsection{\code{List_13F_2015_Q2}}{

A data frame with 17,395 rows and 7 columns
}

\subsection{\code{List_13F_2015_Q3}}{

A data frame with 17,587 rows and 7 columns
}

\subsection{\code{List_13F_2015_Q4}}{

A data frame with 18,013 rows and 7 columns
}

\subsection{\code{List_13F_2016_Q1}}{

A data frame with 17,831 rows and 7 columns
}

\subsection{\code{List_13F_2016_Q2}}{

A data frame with 17,982 rows and 7 columns
}

\subsection{\code{List_13F_2016_Q3}}{

A data frame with 17,776 rows and 7 columns
}

\subsection{\code{List_13F_2016_Q4}}{

A data frame with 17,681 rows and 7 columns
}

\subsection{\code{List_13F_2017_Q1}}{

A data frame with 17,560 rows and 7 columns
}

\subsection{\code{List_13F_2017_Q2}}{

A data frame with 17,675 rows and 7 columns
}

\subsection{\code{List_13F_2017_Q3}}{

A data frame with 17,765 rows and 7 columns
}

\subsection{\code{List_13F_2017_Q4}}{

A data frame with 17,770 rows and 7 columns
}

\subsection{\code{List_13F_2023_Q1}}{

A data frame with 23,988 rows and 7 columns
}

\subsection{\code{List_13F_2023_Q2}}{

A data frame with 23,759 rows and 7 columns
}

\subsection{\code{List_13F_2023_Q3}}{

A data frame with 23,485 rows and 7 columns
}

\subsection{\code{List_13F_2023_Q4}}{

A data frame with 23,425 rows and 7 columns
}

\subsection{\code{List_13F_2024_Q1}}{

A data frame with 23,989 rows and 7 columns
}
}
\source{
\url{https://www.sec.gov/divisions/investment/13flists}
}
\usage{
List_13F_2004_Q1

List_13F_2004_Q2

List_13F_2004_Q3

List_13F_2004_Q4

List_13F_2005_Q1

List_13F_2005_Q2

List_13F_2005_Q3

List_13F_2005_Q4

List_13F_2006_Q1

List_13F_2006_Q2

List_13F_2006_Q3

List_13F_2006_Q4

List_13F_2007_Q1

List_13F_2007_Q2

List_13F_2007_Q3

List_13F_2007_Q4

List_13F_2008_Q1

List_13F_2008_Q2

List_13F_2008_Q3

List_13F_2008_Q4

List_13F_2009_Q1

List_13F_2009_Q2

List_13F_2009_Q3

List_13F_2009_Q4

List_13F_2010_Q1

List_13F_2010_Q2

List_13F_2010_Q3

List_13F_2010_Q4

List_13F_2011_Q1

List_13F_2011_Q2

List_13F_2011_Q3

List_13F_2011_Q4

List_13F_2012_Q1

List_13F_2012_Q2

List_13F_2012_Q3

List_13F_2012_Q4

List_13F_2013_Q1

List_13F_2013_Q2

List_13F_2013_Q3

List_13F_2013_Q4

List_13F_2014_Q1

List_13F_2014_Q2

List_13F_2014_Q3

List_13F_2014_Q4

List_13F_2015_Q1

List_13F_2015_Q2

List_13F_2015_Q3

List_13F_2015_Q4

List_13F_2016_Q1

List_13F_2016_Q2

List_13F_2016_Q3

List_13F_2016_Q4

List_13F_2017_Q1

List_13F_2017_Q2

List_13F_2017_Q3

List_13F_2017_Q4

List_13F_2023_Q1

List_13F_2023_Q2

List_13F_2023_Q3

List_13F_2023_Q4

List_13F_2024_Q1
}
\description{
Parsed data from Official List of Section 13(f) securities.
Data starts from Q1 2004. Data section of the package is updated every quarter,
when new list published on \href{https://www.sec.gov/divisions/investment/13flists}{sec.gov}

Each data set contains the following columns which match data in each PDF list:

\itemize{
\item \code{CUSIP}: character - CUSIP number of the security included in the official list
\item \code{HAS_LISTED_OPTION}: character - An asterisk indicates that security having a listed option and each option is individually listed with its own CUSIP number immediately below the name of the security having the option
\item \code{ISSUER_NAME}: character - Issuer name
\item \code{ISSUER_DESCRIPTION}: character - Issuer description
\item \code{STATUS}: character - "ADDED" (The security has become a Section 13(f) security) or "DELETED" (The security ceases to be a 13(f) security since the date of the last list)
\item \code{YEAR}: integer - Year of the official list
\item \code{QUARTER}: integer - Quarter of the official list
}
}
\keyword{datasets}
