% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_13f_list_local.R
\name{SEC_13F_list_local}
\alias{SEC_13F_list_local}
\title{Official List of Section 13(f) Securities from local file}
\usage{
SEC_13F_list_local(path_to_13f_file)
}
\arguments{
\item{path_to_13f_file}{character string, path to local file}
}
\value{
A data frame that contains official list of Section 13(f) securities with the following columns:
\itemize{
\item \code{CUSIP}: character - CUSIP number of the security included in the official list
\item \code{HAS_LISTED_OPTION}: character - An asterisk indicates that security having a listed option and each option is individually listed with its own CUSIP number immediately below the name of the security having the option
\item \code{ISSUER_NAME}: character - Issuer name
\item \code{ISSUER_DESCRIPTION}: character - Issuer description
\item \code{STATUS}: character - "ADDED" (The security has become a Section 13(f) security) or "DELETED" (The security ceases to be a 13(f) security since the date of the last list)
\item \code{YEAR}: integer - Year of the official list
\item \code{QUARTER}: integer - Quarter of the official list
}
}
\description{
This function specified by Year and Quarter (starting from year 2004, quarter 1), official list of Section 13(f) securities from SEC website <https://www.sec.gov/divisions/investment/13flists.htm>, parses it and returns a data frame. If no parameters provided, function determines year and quarter based on Current List section of SEC website <https://www.sec.gov/divisions/investment/13flists.htm>
}
\examples{
\donttest{library(SEC13Flist)
SEC_13F_list_2018_Q3 <- SEC_13F_list(2018,3) #Parse list for Q3 2018 without progress indicator
SEC_13F_list_2018_Q3_ <- SEC_13F_list(2018,3,TRUE) #Parse list with progress indicator
SEC_13F_list_current <- SEC_13F_list() #Parse current list from SEC.gov
}
}
\keyword{13F}
\keyword{List}
\keyword{SEC}
