% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url_file_func}
\alias{url_file_func}
\title{File URL}
\usage{
url_file_func(YEAR_, QUARTER_, current_year, current_quarter)
}
\arguments{
\item{YEAR_}{year of the list}

\item{QUARTER_}{quarter of the list}

\item{current_year}{current year}

\item{current_quarter}{current quarter}
}
\description{
internal function that determines URL of the list
}
