% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isCusip}
\alias{isCusip}
\title{Check validity of CUSIP code}
\usage{
isCusip(s)
}
\arguments{
\item{s}{nine-character string with CUSIP code to validate}
}
\value{
Logical value indicating validity of CUSIP number: \code{TRUE} if CUSIP number is valid, \code{FALSE} if CUSIP number is invalid
}
\description{
This function check validity of CUSIP code by comparing calculated checksum digit based on first 8 characters of CUSIP code with 9th character of CUSIP code - checksum digit
}
\examples{
library(SEC13Flist)
isCusip("B38564109") #invalid CUSIP example
isCusip("B38564108") #valid CUSIP example
}
\keyword{CUSIP}
\keyword{checksum}
\keyword{digit}
