% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_13f_list.R
\name{SEC_13F_list}
\alias{SEC_13F_list}
\title{Official List of Section 13(f) Securities}
\usage{
SEC_13F_list(YEAR_, QUARTER_)
}
\arguments{
\item{YEAR_}{Year for the SEC List}

\item{QUARTER_}{Quarter for the SEC List}
}
\description{
This function downloads, specified by Year and Quarter, Official List of Section 13(f) Securities from SEC website, parses it and returns dataframe. If no parameters provided, function determines year and quarter based on Current List section of SEC website
}
\examples{
library(SEC13Flist)
SEC_13F_list_2018_Q3 <- SEC_13F_list(2018,3) #Download and parse list for Q3 2018
SEC_13F_list_current <- SEC_13F_list() #Download and parse current list from SEC.gov
SEC13Flist_current <- SEC_13F_list() \%>\%
  filter(STATUS!="DELETED") \%>\% #Filter records with STATUS "DELETED"
  select(-YEAR,-QUARTER) #Remove YEAR and QUARTER columns
}
\keyword{13F}
\keyword{List}
\keyword{SEC}
